+++ release_notes.md
@@ -1,10 +1,12 @@
 Release Notes
 =======
 
-## 3.2.0 (under development)
+## 3.2.0
 **New**
+- Concurrency support - thread-safe access to 'IDomainAccess'.
 - Query DSL and Domain Query - String Operators: STARTS_WITH, CONTAINS (CONTAINS_string), ENDS_WITH
-- Query DSL - Supports returning Literal Maps
+- Query DSL - Support for Literal Maps
+- Upgrade to Neo4j 2.3.1
 
 ## 3.1.0
 **New**